Use entry point for emsymbolizer
